wip Settings fragments
2020-04-05 23:58:23.548 27589-27589/org.runnerup.debug E/AndroidRuntime: FATAL EXCEPTION: main
    Process: org.runnerup.debug, PID: 27589
    java.lang.StackOverflowError: stack size 8MB
        at android.content.res.AssetManager.getThemeValue(AssetManager.java:644)
        at android.content.res.ResourcesImpl$ThemeImpl.resolveAttribute(ResourcesImpl.java:1374)
        at android.content.res.Resources$Theme.resolveAttribute(Resources.java:1578)
        at androidx.preference.PreferenceFragmentCompat.onCreate(PreferenceFragmentCompat.java:143)
        at org.runnerup.view.fragments.PrefsFragment.onCreatePreferences(PrefsFragment.java:57)
        at androidx.preference.PreferenceFragmentCompat.onCreate(PreferenceFragmentCompat.java:160)
        at org.runnerup.view.fragments.PrefsFragment.onCreatePreferences(PrefsFragment.java:57)
        at androidx.preference.PreferenceFragmentCompat.onCreate(PreferenceFragmentCompat.java:160)
